+++ Parser.py
@@ -1,0 +1,6 @@
+class Token:
+
+    def __init__(self, _type, value):
+        self.type = type(_type)
+        self.value = value
+
